requirements.txt
python-dateutil==2.8.0
requests==2.22.0
requests-toolbelt==0.9.1
texttable==1.6.1
urllib3==1.25.7
